Very first working version of ECB_collector
this is the first version of the raw ECB csv. It is dogshit for the following reasons:

- Output is completely unfiltered, just a huge list of every character and move
- LibMelee names for special action states are very stupid and should not be trusted
- Doesn't collect data on some of the most important ECBs: jump, DJ, airdodge, and moves with directionally-dependent ECBs (e.g., firefox)
- Intended to uns with zero-gravity gecko code over BF side platform and UnclePunch's LRA-Start QoL codes, so the many specials that induce gravity/downward motion only get their first few frames then subsequent landing captured. Without these codes, no guarantee this shit works.
- Didn't run it for Mewtwo neutral-b, it fucks up on certain Ness sequences, and ignores anything of Sheik, and it runs twice for Ice Climbers, since these issues would've required special changes to my program and I wanted to get a first version done

Still, I think this is the first complete program to collect the ECBs of nearly all normal aerials in the game. Every issue above is fixable with some tinkering and time investment, so I think the odds are good that I can publish a much better, cleaner version of this. In case I never get that done, I wanted to put this out there 🙂

GuyDino THUNDER godspeed